Add files via upload
Pharmacy Inventory System project is written in C++ programming language. Talking about the system, it contains all the required functions which include adding, viewing, deleting, and updating medicine items.
 It is a C++ object-oriented-based desktop application to manage the storage of medicines. 
This system served as an administrative programmed and users (Admin) need to key in a password to login into the system. 
This system allows users to add, update, delete, view and search the medical records according to the medicine id number